+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
 
 * Remove `implants` from `EveOnline::ESI::Models::JumpClone#as_json`
+* Rename `EveOnline::ESI::Models::JumpClone#implants` to `#implant_ids`
 
 ## v0.23.0